Rename macro tests + update directories
