+++ fuel-indexer-macros/src/lib.rs
@@ -21,13 +21,19 @@ mod tests {
         let t = trybuild::TestCases::new();
         std::env::set_var("COMPILE_TEST_PREFIX", env!("CARGO_MANIFEST_DIR"));
 
-        t.compile_fail("test_data/fail_self.rs");
-        t.compile_fail("test_data/fail_args.rs");
-        t.compile_fail("test_data/fail_args2.rs");
-        t.pass("test_data/success.rs");
-        t.pass("test_data/success2.rs");
-        t.compile_fail("test_data/fail_badschema.rs");
-        t.compile_fail("test_data/fail_unknown_type.rs");
-        t.compile_fail("test_data/fail_empty.rs");
+        t.compile_fail("./../../tests/assets/macro-data/fail_if_attribute_args_include_self.rs");
+        t.compile_fail("./../../tests/assets/macro-data/fail_if_attribute_args_not_included.rs");
+        t.compile_fail(
+            "./../../tests/assets/macro-data/fail_if_all_attribute_args_not_included.rs",
+        );
+        t.pass("./../../tests/assets/macro-data/pass_if_indexer_is_valid_single_type.rs");
+        t.pass("./../../tests/assets/macro-data/pass_if_indexer_is_valid_multi_type.rs");
+        t.compile_fail(
+            "./../../tests/assets/macro-data/fail_if_attribute_schema_arg_is_invalid.rs",
+        );
+        t.compile_fail(
+            "./../../tests/assets/macro-data/fail_if_attribute_abi_arg_includes_invalid_type.rs",
+        );
+        t.compile_fail("./../../tests/assets/macro-data/fail_if_indexer_module_is_empty.rs");
     }
 }
