--- fuel-indexer-macros/test_data/contracts-abi.json
@@ -1,74 +1,0 @@
-[
-  {
-    "type": "function",
-    "inputs": [
-      {
-        "name": "num",
-        "type": "u64",
-        "components": null,
-        "typeArguments": null
-      }
-    ],
-    "name": "gimme_someevent",
-    "outputs": [
-      {
-        "name": "",
-        "type": "struct SomeEvent",
-        "components": [
-          {
-            "name": "id",
-            "type": "u64",
-            "components": null,
-            "typeArguments": null
-          },
-          {
-            "name": "account",
-            "type": "b256",
-            "components": null,
-            "typeArguments": null
-          }
-        ],
-        "typeArguments": null
-      }
-    ]
-  },
-  {
-    "type": "function",
-    "inputs": [
-      {
-        "name": "num",
-        "type": "u64",
-        "components": null,
-        "typeArguments": null
-      }
-    ],
-    "name": "gimme_anotherevent",
-    "outputs": [
-      {
-        "name": "",
-        "type": "struct AnotherEvent",
-        "components": [
-          {
-            "name": "id",
-            "type": "u64",
-            "components": null,
-            "typeArguments": null
-          },
-          {
-            "name": "account",
-            "type": "b256",
-            "components": null,
-            "typeArguments": null
-          },
-          {
-            "name": "hash",
-            "type": "b256",
-            "components": null,
-            "typeArguments": null
-          }
-        ],
-        "typeArguments": null
-      }
-    ]
-  }
-]
