+++ tests/assets/compose-up.bash
