+++ tests/assets/contracts/ping/Forc.toml
