+++ tests/assets/contracts/ping/src/main.sw
