+++ tests/assets/contracts/simple_wasm/.gitignore
