+++ tests/assets/contracts/simple_wasm/Forc.toml
