+++ tests/assets/contracts/simple_wasm/src/main.sw
