+++ tests/assets/docker-compose.yaml
