+++ tests/assets/macro-data/fail_if_attribute_abi_arg_includes_invalid_type.rs
