+++ tests/assets/macro-data/fail_if_attribute_args_include_self.stderr
