+++ tests/assets/macro-data/fail_if_attribute_args_not_included.rs
