+++ tests/assets/macro-data/fail_if_attribute_schema_arg_is_invalid.stderr
