+++ tests/assets/macro-data/fail_if_indexer_module_is_empty.rs
