+++ tests/assets/macro-data/pass_if_indexer_is_valid_multi_type.rs
