+++ tests/assets/macro-data/pass_if_indexer_is_valid_single_type.rs
@@ -6,10 +6,10 @@ fn ff_log_data(_inp: ()) {}
 
 
 #[indexer(
-    abi = "./test_data/contracts-abi.json",
+    abi = "./../contracts/simple_wasm/out/debug/contracts-abi.json",
     namespace = "test_namespace",
     identifier = "index1",
-    schema = "./test_data/schema.graphql",
+    schema = "./../schema/schema.graphql",
 )]
 mod indexer {
     fn function_one(event: SomeEvent) {
