+++ tests/assets/manifest/composability_test.yaml
