+++ tests/assets/schema/bad_simple_wasm.graphql
