+++ tests/assets/schema/composability_test.graphql
