+++ tests/assets/schema/simple_wasm.graphql
